app-updated

+++ src/Forms/Headquarters.jsx
@@ -4,6 +4,7 @@ import { collection, addDoc, getDocs, deleteDoc, doc, updateDoc } from "firebase
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import './Headquarters.css'; // Import separate CSS file
+import { useNavigate } from 'react-router-dom';
 
 function Headquarters() {
   const [place, setPlace] = useState('');
@@ -12,9 +13,13 @@ function Headquarters() {
   const [headquarters, setHeadquarters] = useState([]);
   const [loading, setLoading] = useState(false);
   const [showAddForm, setShowAddForm] = useState(false);
+  const navigate=useNavigate()
   const [searchTerm, setSearchTerm] = useState('');
-  const [editingHeadquarter, setEditingHeadquarter] = useState(null); // Track which headquarter is being edited
-  const headquartersPerPage = 10;
+  const [editingHeadquarter, setEditingHeadquarter] = useState(null);
+  const [showDeleteConfirmation, setShowDeleteConfirmation] = useState(false);
+  const [deletePin, setDeletePin] = useState('');
+  const [headquarterToDelete, setHeadquarterToDelete] = useState(null); // Track which headquarter is being deleted
+  const headquartersPerPage = 20;
   const [currentPage, setCurrentPage] = useState(1);
 
   useEffect(() => {
@@ -70,13 +75,21 @@ function Headquarters() {
   };
 
   const handleDeleteHeadquarter = async (id) => {
-    try {
-      await deleteDoc(doc(db, "Headquarters", id));
-      toast.success('Headquarter deleted successfully!');
-      fetchHeadquarters();
-    } catch (error) {
-      console.error("Error deleting document: ", error);
-      toast.error('Error deleting headquarter. Please try again.');
+    if (deletePin === '2024') {
+      try {
+        await deleteDoc(doc(db, "Headquarters", id));
+        toast.success('Headquarter deleted successfully!');
+        fetchHeadquarters();
+      } catch (error) {
+        console.error("Error deleting document: ", error);
+        toast.error('Error deleting headquarter. Please try again.');
+      } finally {
+        setShowDeleteConfirmation(false);
+        setDeletePin('');
+        setHeadquarterToDelete(null);
+      }
+    } else {
+      toast.error('Incorrect PIN. Please try again.');
     }
   };
 
@@ -141,8 +154,44 @@ function Headquarters() {
         pauseOnHover
       />
 
+      {/* Delete Confirmation Modal */}
+      {showDeleteConfirmation && (
+        <div className="delete-confirmation-modal">
+          <div className="delete-confirmation-content">
+            <h3>Confirm Deletion</h3>
+            <p>Enter the PIN "2024" to confirm deletion:</p>
+            <input
+              type="text"
+              value={deletePin}
+              onChange={(e) => setDeletePin(e.target.value)}
+              placeholder="Enter PIN"
+              className="delete-pin-input"
+            />
+            <div className="delete-confirmation-buttons">
+              <button
+                className="delete-confirm-button"
+                onClick={() => handleDeleteHeadquarter(headquarterToDelete)}
+              >
+                Confirm
+              </button>
+              <button
+                className="delete-cancel-button"
+                onClick={() => {
+                  setShowDeleteConfirmation(false);
+                  setDeletePin('');
+                  setHeadquarterToDelete(null);
+                }}
+              >
+                Cancel
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
+        <button onClick={() => navigate(-1)} className="back-button" style={{color:"#d6e8ee"}}><i className="bi bi-arrow-left"></i></button>
+
       <div className="headquarters-header">
-        <h1 className="headquarters-title">Headquarters Management</h1>
+        <h1 className="headquarters-title">Headquarters</h1>
         <button className="headquarters-add-button" onClick={() => setShowAddForm(!showAddForm)}>
           {showAddForm ? 'Hide Form' : 'Add New Headquarters'}
         </button>
@@ -168,7 +217,6 @@ function Headquarters() {
               id="location"
               value={location}
               onChange={(e) => setLocation(e.target.value)}
-              required
               className="headquarters-input"
             />
           </div>
@@ -183,6 +231,7 @@ function Headquarters() {
                   onChange={(e) => handleStaffChange(index, 'name', e.target.value)}
                   className="staff-input"
                 />
+                <div className="div-hide-date">
                 <input
                   type="text"
                   placeholder="Address"
@@ -190,6 +239,7 @@ function Headquarters() {
                   onChange={(e) => handleStaffChange(index, 'address', e.target.value)}
                   className="staff-input"
                 />
+                </div>
                 <input
                   type="text"
                   placeholder="Phone Number"
@@ -197,13 +247,16 @@ function Headquarters() {
                   onChange={(e) => handleStaffChange(index, 'phoneNumber', e.target.value)}
                   className="staff-input"
                 />
+                <div className="div-hide-date">
                 <input
                   type="email"
                   placeholder="Email"
+                  
                   value={staffMember.email}
                   onChange={(e) => handleStaffChange(index, 'email', e.target.value)}
                   className="staff-input"
                 />
+                </div>
               </div>
             ))}
             <button type="button" className="add-staff-button" onClick={handleAddStaff}>
@@ -211,7 +264,7 @@ function Headquarters() {
             </button>
           </div>
           <button type="submit" className="headquarters-submit-button" disabled={loading}>
-            {loading ? 'Adding...' : 'Add Headquarters'}
+            {loading ? 'Adding...' : 'Save'}
           </button>
         </form>
       )}
@@ -226,119 +279,130 @@ function Headquarters() {
         />
       </div>
 
-      <table className="headquarters-table">
-        <thead>
-          <tr>
-            <th>Place</th>
-            <th>Location</th>
-            <th>Staff</th>
-            <th>Actions</th>
-          </tr>
-        </thead>
-        <tbody>
-          {currentHeadquarters.map(headquarter => (
-            <tr key={headquarter.id}>
-              <td>
-                {editingHeadquarter?.id === headquarter.id ? (
-                  <input
-                    type="text"
-                    value={editingHeadquarter.place}
-                    onChange={(e) =>
-                      setEditingHeadquarter({ ...editingHeadquarter, place: e.target.value })
-                    }
-                  />
-                ) : (
-                  headquarter.place
-                )}
-              </td>
-              <td>
-                {editingHeadquarter?.id === headquarter.id ? (
-                  <input
-                    type="text"
-                    value={editingHeadquarter.location}
-                    onChange={(e) =>
-                      setEditingHeadquarter({ ...editingHeadquarter, location: e.target.value })
-                    }
-                  />
-                ) : (
-                  headquarter.location
-                )}
-              </td>
-              <td>
-                {editingHeadquarter?.id === headquarter.id ? (
-                  <>
-                    {editingHeadquarter.staff.map((staffMember, index) => (
-                      <div key={index} className="staff-form-group">
-                        <input
-                          type="text"
-                          placeholder="Name"
-                          value={staffMember.name}
-                          onChange={(e) => handleEditStaff(editingHeadquarter, index, 'name', e.target.value)}
-                          className="staff-input"
-                        />
-                        <input
-                          type="text"
-                          placeholder="Address"
-                          value={staffMember.address}
-                          onChange={(e) => handleEditStaff(editingHeadquarter, index, 'address', e.target.value)}
-                          className="staff-input"
-                        />
-                        <input
-                          type="text"
-                          placeholder="Phone Number"
-                          value={staffMember.phoneNumber}
-                          onChange={(e) => handleEditStaff(editingHeadquarter, index, 'phoneNumber', e.target.value)}
-                          className="staff-input"
-                        />
-                        <input
-                          type="email"
-                          placeholder="Email"
-                          value={staffMember.email}
-                          onChange={(e) => handleEditStaff(editingHeadquarter, index, 'email', e.target.value)}
-                          className="staff-input"
-                        />
-                      </div>
-                    ))}
-                    <button type="button" className="add-staff-button" onClick={() => handleAddStaffInEdit(editingHeadquarter)}>
-                      Add More Staff
-                    </button>
-                  </>
-                ) : (
-                  <ul>
-                  {(headquarter.staff || []).map((staffMember, index) => (
-                    <li key={index}>
-                      <strong>{staffMember.name}</strong> - {staffMember.address}, {staffMember.phoneNumber}, {staffMember.email}
-                    </li>
-                  ))}
-                </ul>
-
-                )}
-              </td>
-              <td>
-                {editingHeadquarter?.id === headquarter.id ? (
-                  <>
-                    <button className="headquarters-save-button" onClick={() => handleUpdateHeadquarter(editingHeadquarter)}>
-                      Save
-                    </button>
-                    <button className="headquarters-cancel-button" onClick={() => setEditingHeadquarter(null)}>
-                      Cancel
-                    </button>
-                  </>
-                ) : (
-                  <>
-                    <button className="headquarters-update-button" onClick={() => setEditingHeadquarter(headquarter)}>
-                      Update
-                    </button>
-                    <button className="headquarters-delete-button" onClick={() => handleDeleteHeadquarter(headquarter.id)}>
-                      Delete
-                    </button>
-                  </>
-                )}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
+      <div className="table-wrapper">
+  <table className="headquarters-table">
+    <thead>
+      <tr>
+        <th>Place</th>
+        <th>Location</th>
+        <th>Staff</th>
+        <th>Actions</th>
+      </tr>
+    </thead>
+    <tbody>
+      {currentHeadquarters.map(headquarter => (
+        <tr key={headquarter.id}>
+          <td>
+            {editingHeadquarter?.id === headquarter.id ? (
+              <input
+                type="text"
+                value={editingHeadquarter.place}
+                onChange={(e) =>
+                  setEditingHeadquarter({ ...editingHeadquarter, place: e.target.value })
+                }
+              />
+            ) : (
+              headquarter.place
+            )}
+          </td>
+          <td>
+            {editingHeadquarter?.id === headquarter.id ? (
+              <input
+                type="text"
+                value={editingHeadquarter.location}
+                onChange={(e) =>
+                  setEditingHeadquarter({ ...editingHeadquarter, location: e.target.value })
+                }
+              />
+            ) : (
+              headquarter.location
+            )}
+          </td>
+          <td>
+            {editingHeadquarter?.id === headquarter.id ? (
+              <>
+                {editingHeadquarter.staff.map((staffMember, index) => (
+                  <div key={index} className="staff-form-group">
+                    <input
+                      type="text"
+                      placeholder="Name"
+                      value={staffMember.name}
+                      onChange={(e) => handleEditStaff(editingHeadquarter, index, 'name', e.target.value)}
+                      className="staff-input"
+                    />
+                    <div className="div-hide-date">
+                    <input
+                      type="text"
+                      placeholder="Address"
+                      value={staffMember.address}
+                      onChange={(e) => handleEditStaff(editingHeadquarter, index, 'address', e.target.value)}
+                      className="staff-input"
+                    />
+                    </div>
+                    <input
+                      type="text"
+                      placeholder="Phone Number"
+                      value={staffMember.phoneNumber}
+                      onChange={(e) => handleEditStaff(editingHeadquarter, index, 'phoneNumber', e.target.value)}
+                      className="staff-input"
+                    />
+                    <div className="div-hide-date">
+                    <input
+                      type="email"
+                      placeholder="Email"
+                      value={staffMember.email}
+                      onChange={(e) => handleEditStaff(editingHeadquarter, index, 'email', e.target.value)}
+                      className="staff-input"
+                    />
+                    </div>
+                  </div>
+                ))}
+                <button type="button" className="add-staff-button" onClick={() => handleAddStaffInEdit(editingHeadquarter)}>
+                  Add More Staff
+                </button>
+              </>
+            ) : (
+              <ul>
+                {(headquarter.staff || []).map((staffMember, index) => (
+                  <li key={index}>
+                    <strong>{staffMember.name}</strong>, {staffMember.phoneNumber}
+                  </li>
+                ))}
+              </ul>
+            )}
+          </td>
+          <td>
+            {editingHeadquarter?.id === headquarter.id ? (
+              <>
+                <button className="headquarters-save-button" onClick={() => handleUpdateHeadquarter(editingHeadquarter)}>
+                  Save
+                </button>
+                <button className="headquarters-cancel-button" onClick={() => setEditingHeadquarter(null)}>
+                  Cancel
+                </button>
+              </>
+            ) : (
+              <>
+                <button className="headquarters-update-button btn btn-success" onClick={() => setEditingHeadquarter(headquarter)}>
+                  Update
+                </button>
+                <button
+                  className="headquarters-delete-button btn btn-danger"
+                  onClick={() => {
+                    setHeadquarterToDelete(headquarter.id);
+                    setShowDeleteConfirmation(true);
+                  }}
+                >
+                  Delete
+                </button>
+              </>
+            )}
+          </td>
+        </tr>
+      ))}
+    </tbody>
+  </table>
+</div>
 
       <div className="headquarters-pagination">
         <button
